+++ packages/gamut/src/ConnectedForm/index.tsx
@@ -1,4 +1,5 @@
-export * from '../ConnectedForm/FormWrapper';
-export * from '../ConnectedForm/SubmitButton';
-export * from '../ConnectedForm/ConnectedFormGroup';
-export * from '../ConnectedForm/utils';
+export * from './ConnectedFormGroup';
+export * from './FormWrapper';
+export * from './Inputs';
+export * from './SubmitButton';
+export * from './utils';
