--- packages/styleguide/stories/Organisms/Modular Forms/examples.tsx
@@ -1,15 +1,0 @@
-import {
-  ConnectedCheckbox,
-  FormWrapper,
-  SubmitButton,
-} from '@codecademy/gamut/src';
-import React from 'react';
-
-export const HEY = () => {
-  return (
-    <FormWrapper onSubmit={() => console.log('hey')}>
-      <ConnectedCheckbox namf="fdkfsldk" />
-      <SubmitButton>submit!!!</SubmitButton>
-    </FormWrapper>
-  );
-};
